+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### App Center
 
-* **[Feature]** Catch "low memory warning" and provide the API to check if it has happened in last session:  `Crashes.hasReceivedMemoryWarningInLastSession()`.
+* **[Feature]** Catch low memory warning and provide the API to check if it has happened in last session:  `Crashes.hasReceivedMemoryWarningInLastSession()`.
 
 ## Version 2.2.0